data pipeline/requirements.txt
paho-mqtt>=1.6.1
Flask>=2.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
requests>=2.28
influxdb>=5.3.1
psycopg2-binary>=2.9.6
pydantic>=2.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
APScheduler>=3.10.0
